Move from_chars floating implementation into src folder
